refactor: Singularize method names
BREAKING CHANGE: The trigger model is now singular, update your import 
statements
