+++ packages/cozy-client/src/models/index.js
@@ -1,8 +1,8 @@
-import * as triggers from './triggers'
+import * as trigger from './trigger'
 import * as instance from './instance'
 import * as applications from './applications'
 import * as file from './file'
-import * as accounts from './accounts'
+import * as account from './account'
 import * as note from './note'
 
-export { triggers, instance, applications, file, note, accounts }
+export { trigger, instance, applications, file, note, account }
